fix(trellis): explicit double casts in L2-norm sums (-Werror=double-promotion)
The two L2-norm accumulator loops in ggml_trellis_encode_group cast only
the left operand to double, leaving the right operand as float. GCC and
Clang under -Werror=double-promotion (used by upstream CI) reject this as
an implicit conversion. Add the matching cast on the right operand to
make the intent explicit. Numerically identical.

+++ ggml/src/ggml-trellis.c
@@ -171,7 +171,7 @@ void ggml_trellis_encode_group(
 
     // Group L2 norm
     double n2 = 0.0;
-    for (int j = 0; j < N; j++) n2 += (double)x[j] * x[j];
+    for (int j = 0; j < N; j++) n2 += (double)x[j] * (double)x[j];
     float norm = (float)sqrt(n2);
     if (norm < 1e-30f) {
         *out_start_state = 0;
@@ -309,7 +309,7 @@ void ggml_trellis_encode_group(
     double recon_sq = 0.0;
     for (int i = 0; i < N; i++) {
         float code = table[states[i + 1]] * cb_scale;
-        recon_sq += (double)code * code;
+        recon_sq += (double)code * (double)code;
     }
     float recon_norm = (float)sqrt(recon_sq);
     *out_d = (recon_norm > 1e-30f) ? (norm / recon_norm) : norm;
